Move CiftiSurfaceResample from utils to interfaces.
